requirements.txt
# Core ML/AI dependencies
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
timm>=0.9.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine learning utilities
scikit-learn>=1.3.0
torchmetrics>=0.11.0

# API and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Google AI integration
google-generativeai>=0.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Jupyter and notebook support
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Logging and monitoring
loguru>=0.7.0
tqdm>=4.65.0

# Image processing
albumentations>=1.3.0

# Optional: for advanced visualization
tensorboard>=2.13.0
wandb>=0.15.0